Rename borrrowed to model
